Factor Contract Summary out to its own component
